bunch of moving things around and feature-locking
