+++ common/cosmwasm-smart-contracts/testing/Cargo.toml
@@ -6,15 +6,20 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-base64 = "0.21.0"
+base64 = { version = "0.21.0", optional = true }
 cosmwasm-std = { version = "1.0.0" }
-cosmwasm-storage = { version = "1.0.0" }
+cosmwasm-storage = { version = "1.0.0", optional = true }
 cw-storage-plus = { version = "0.13.4", optional = true }
-hex = "0.4.3"
-rand_chacha = "0.3"
+hex = { version = "0.4.3", optional = true }
+rand_chacha = { version = "0.3", optional = true }
 serde = { version = "1", features=["derive"] }
-serde_json = "1"
-thiserror = "1.0.38"
+serde_json = { version = "1", optional = true }
+thiserror = { version = "1.0.38" }
 
 [features]
-default = ["cw-storage-plus"]
+default = ["full"]
+full = ["testable-trait", "contract-mocks", "rand", "state-importing", "cosmwasm-storage", "cw-storage-plus", "serde_json"]
+rand = ["rand_chacha"]
+contract-mocks = []
+state-importing = ["base64", "hex"]
+testable-trait = []
