+++ contracts/vesting/src/support/tests.rs
@@ -1,104 +1,24 @@
 #[cfg(test)]
 pub mod helpers {
-
-    // TODO: once https://github.com/nymtech/nym/pull/3040 gets merged,
-    // the `ContractState` should replace the below
-    #[allow(unused)]
-    mod state_dump_decoder {
-        use base64::{engine::general_purpose, Engine};
-        use serde::{Deserialize, Serialize};
-        use std::fs::File;
-        use std::path::Path;
-
-        #[derive(Serialize, Deserialize, Debug)]
-        pub struct RawState {
-            pub height: String,
-            pub result: Vec<RawKV>,
-        }
-
-        impl RawState {
-            pub fn decode(self) -> DecodedState {
-                DecodedState {
-                    height: self.height.parse().unwrap(),
-                    result: self
-                        .result
-                        .into_iter()
-                        .map(|raw| DecodedKV {
-                            key: hex::decode(&raw.key).unwrap(),
-                            value: general_purpose::STANDARD.decode(&raw.value).unwrap(),
-                        })
-                        .collect(),
-                }
-            }
-
-            pub fn from_file<P: AsRef<Path>>(path: P) -> Self {
-                let file = File::open(path).expect("failed to open specified file");
-                serde_json::from_reader(file).expect("failed to parse specified file")
-            }
-        }
-
-        #[derive(Serialize, Deserialize, Debug)]
-        pub struct RawKV {
-            // hex
-            pub key: String,
-
-            // base64
-            pub value: String,
-        }
-
-        #[derive(Serialize, Deserialize, Debug)]
-        pub struct DecodedKV {
-            pub key: Vec<u8>,
-            pub value: Vec<u8>,
-        }
-
-        #[derive(Serialize, Deserialize, Debug)]
-        pub struct DecodedState {
-            pub height: u64,
-            pub result: Vec<DecodedKV>,
-        }
-
-        impl DecodedState {
-            pub fn find_value(&self, key: &[u8]) -> Option<Vec<u8>> {
-                self.result
-                    .iter()
-                    .find(|kv| kv.key == key)
-                    .map(|kv| kv.value.clone())
-            }
-        }
-    }
-
     use crate::contract::{instantiate, try_create_periodic_vesting_account};
     use crate::storage::{ACCOUNTS, ADMIN, MIXNET_CONTRACT_ADDRESS, MIX_DENOM};
-    use crate::support::tests::helpers::state_dump_decoder::RawState;
     use crate::traits::VestingAccount;
     use crate::vesting::{populate_vesting_periods, Account};
     use contracts_common::Percent;
+    use cosmwasm_contract_testing::{env_with_block_info, ContractState};
     use cosmwasm_std::testing::{mock_dependencies, mock_env, mock_info, MockApi, MockQuerier};
     use cosmwasm_std::{
-        coin, Addr, BlockInfo, Coin, ContractInfo, Deps, DepsMut, Empty, Env, MemoryStorage,
-        MessageInfo, OwnedDeps, Storage, Timestamp, Uint128,
+        coin, Addr, BlockInfo, Coin, Deps, DepsMut, Empty, Env, MemoryStorage, MessageInfo,
+        OwnedDeps, Storage, Timestamp, Uint128,
     };
-    use rand_chacha::rand_core::SeedableRng;
-    use rand_chacha::ChaCha20Rng;
     use std::path::Path;
     use std::str::FromStr;
     use vesting_contract_common::messages::{InitMsg, VestingSpecification};
     use vesting_contract_common::PledgeCap;
 
-    // use rng with constant seed for all tests so that they would be deterministic
-    #[allow(unused)]
-    pub fn test_rng() -> ChaCha20Rng {
-        let dummy_seed = [42u8; 32];
-        rand_chacha::ChaCha20Rng::from_seed(dummy_seed)
-    }
-
     #[allow(unused)]
     pub struct TestSetup {
-        pub deps: OwnedDeps<MemoryStorage, MockApi, MockQuerier<Empty>>,
-        pub env: Env,
-        pub rng: ChaCha20Rng,
-
+        pub state: ContractState,
         pub admin: MessageInfo,
     }
 
@@ -109,40 +29,28 @@ pub mod helpers {
             let admin = ADMIN.load(deps.as_ref().storage).unwrap();
 
             TestSetup {
-                deps,
-                env: mock_env(),
-                rng: test_rng(),
+                state: ContractState::new(),
                 admin: mock_info(admin.as_str(), &[]),
             }
         }
 
         pub fn new_from_state_dump<P: AsRef<Path>>(dump_file: P) -> Self {
-            let state = RawState::from_file(dump_file).decode();
-
-            let mut deps = mock_dependencies();
-            for kv in state.result {
-                deps.storage.set(&kv.key, &kv.value)
-            }
-
-            let admin = ADMIN.load(deps.as_ref().storage).unwrap();
-            let env = Env {
-                block: BlockInfo {
-                    height: 5633424,
-                    time: Timestamp::from_seconds(1676025955),
-                    chain_id: "nyx".to_string(),
-                },
-                transaction: None,
-                contract: ContractInfo {
-                    address: Addr::unchecked(
-                        "n1nc5tatafv6eyq7llkr2gv50ff9e22mnf70qgjlv737ktmt4eswrq73f2nw",
-                    ),
-                },
+            let current_block = BlockInfo {
+                height: 5633424,
+                time: Timestamp::from_seconds(1676025955),
+                chain_id: "nyx".to_string(),
             };
+            let custom_env = env_with_block_info(current_block);
+            let state = ContractState::try_from_state_dump(dump_file, Some(custom_env.clone()))
+                .unwrap()
+                .with_contract_address(
+                    "n1nc5tatafv6eyq7llkr2gv50ff9e22mnf70qgjlv737ktmt4eswrq73f2nw",
+                );
+
+            let admin = ADMIN.load(state.deps().storage).unwrap();
 
             TestSetup {
-                deps,
-                env,
-                rng: test_rng(),
+                state,
                 admin: mock_info(admin.as_str(), &[]),
             }
         }
@@ -173,15 +81,15 @@ pub mod helpers {
         }
 
         pub fn deps(&self) -> Deps<'_> {
-            self.deps.as_ref()
+            self.state.deps()
         }
 
         pub fn deps_mut(&mut self) -> DepsMut<'_> {
-            self.deps.as_mut()
+            self.state.deps_mut()
         }
 
         pub fn env(&self) -> Env {
-            self.env.clone()
+            self.state.env_cloned()
         }
 
         pub fn admin(&self) -> MessageInfo {
@@ -223,18 +131,18 @@ pub mod helpers {
 
             let pretty = format!(
                 r#"
-{:<20}{original}
-{:<20}{vesting}
-{:<20}{vested}
-{:<20}{balance}
-{:<20}{withdrawn}
-{:<20}{historical_rewards}
-{:<20}{locked}
-{:<20}{spendable}
-{:<20}{spendable_vested}
-{:<20}{spendable_reward}
-{:<20}{total_delegated}
-"#,
+        {:<20}{original}
+        {:<20}{vesting}
+        {:<20}{vested}
+        {:<20}{balance}
+        {:<20}{withdrawn}
+        {:<20}{historical_rewards}
+        {:<20}{locked}
+        {:<20}{spendable}
+        {:<20}{spendable_vested}
+        {:<20}{spendable_reward}
+        {:<20}{total_delegated}
+        "#,
                 "original",
                 "vesting",
                 "vested:",
@@ -265,37 +173,37 @@ pub mod helpers {
 
         pub fn vesting_coins(&self, account: &Account) -> Coin {
             account
-                .get_vesting_coins(None, &self.env, self.deps().storage)
+                .get_vesting_coins(None, self.state.env(), self.deps().storage)
                 .unwrap()
         }
 
         pub fn vested_coins(&self, account: &Account) -> Coin {
             account
-                .get_vested_coins(None, &self.env, self.deps().storage)
+                .get_vested_coins(None, self.state.env(), self.deps().storage)
                 .unwrap()
         }
 
         pub fn locked_coins(&self, account: &Account) -> Coin {
             account
-                .locked_coins(None, &self.env, self.deps().storage)
+                .locked_coins(None, self.state.env(), self.deps().storage)
                 .unwrap()
         }
 
         pub fn spendable_coins(&self, account: &Account) -> Coin {
             account
-                .spendable_coins(None, &self.env, self.deps().storage)
+                .spendable_coins(None, self.state.env(), self.deps().storage)
                 .unwrap()
         }
 
         pub fn spendable_vested_coins(&self, account: &Account) -> Coin {
             account
-                .spendable_vested_coins(None, &self.env, self.deps().storage)
+                .spendable_vested_coins(None, self.state.env(), self.deps().storage)
                 .unwrap()
         }
 
         pub fn spendable_reward_coins(&self, account: &Account) -> Coin {
             account
-                .spendable_reward_coins(None, &self.env, self.deps().storage)
+                .spendable_reward_coins(None, self.state.env(), self.deps().storage)
                 .unwrap()
         }
